+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+- Updates the PR Changelog requirement to include complete descriptions of updates and a link to the
+  associated PR. [PR 572](https://github.com/NatLabRockies/H2Integrate/pull/572)
+- Removed unnecessary `tech_name` designations for some control techs in yamls
+- PySAM battery now takes in charge rate and storage capacity as inputs[PR 557](https://github.com/NatLabRockies/H2Integrate/pull/557)
+
 ## 0.7 [March 3, 2026]
 
 ### New Features